requirements-dev.txt
-r requirements.txt
mypy
pre-commit
ruff
scipy-stubs
specialist
ty
